# swagger.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Work Manager
  description: >-
    Customizable work item manager that leverages Camunda workflows (BPMN and DMN), Form.io and customizable data types
    to model online transaction processing.
servers:
  - url: /api/v1
paths:
...
  /records:
    post:
      summary: Create a new record instance.
      operationId: postRecord
      tags: [ 'Records' ]
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        description: Record Creation Request.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCreationRequest'
      responses:
        200:
          description: The posted Record post-save. This will include assigned IDs for newly created record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponseModel'
    get:
      summary: Get filtered list of records
      operationId: getRecords
      tags: [ 'Records' ]
      parameters:
        - in: query
          name: recordDefinitionKey
          description: Key of the record definition to fetch records for.
          required: false
          schema:
            type: string
        - in: query
          name: status
          description: status to filter records by.
          required: false
          schema:
            type: array
            items:
              type: string
        - in: query
          name: externalId
          description: A external id to fetch records for.
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/sortByRecord"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/pageNumber"
        - $ref: "#/components/parameters/pageSize"
      responses:
        '200':
          description: Collection of records matching filter criteria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedRecordResponseModel'
        "403":
          description: "You are unauthorized to make this request"
        "404":
          description: "Resource not found"
  /records/{id}:
    get:
      summary: Get a single record by ID.
      operationId: getRecord
      tags: [ 'Records' ]
      parameters:
        - in: path
          name: id
          description: ID of record to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponseModel'
        '404':
          description: A record with the given ID does not exist.
    put:
      summary: Update an existing record.
      operationId: updateRecord
      tags: [ 'Records' ]
      parameters:
        - in: path
          name: id
          description: ID of record to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Record Data to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordUpdateRequest'
      responses:
        '200':
          description: The updated record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponseModel'
...
components:
  schemas:
...
    PagedRecordResponseModel:
      description: Paged response detailing a collection of records.
      type: object
      required:
        - items
        - pagingMetadata
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecordResponseModel'
        pagingMetadata:
          $ref: '#/components/schemas/PagingMetadata'
...
    RecordCreationRequest:
      type: object
      required:
        - recordDefinitionKey
        - transactionId
      properties:
        recordDefinitionKey:
          type: string
          pattern: '^[a-zA-Z0-9]+$'
        transactionId:
          type: string
          format: uuid
    RecordUpdateRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          additionalProperties: true
        status:
          type: string
        expires:
          type: string
          format: date-time
    RecordResponseModel:
      type: object
      required:
        - data
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        recordDefinitionId:
          type: string
          format: uuid
        recordDefinitionKey:
          type: string
        recordDefinitionName:
          type: string
        externalId:
          type: string
        createdBy:
          type: string
        createdFrom:
          type: string
          format: uuid
        lastUpdatedBy:
          type: string
        createdTimestamp:
          type: string
          format: date-time
        expires:
          type: string
          format: date-time
        lastUpdatedFrom:
          type: string
          format: uuid
        lastUpdatedTimestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
...
    PagingMetadata:
      description: Paging metadata in accordance with https://datatracker.ietf.org/doc/html/rfc8977
      type: object
      properties:
        totalCount:
          type: integer
          format: int64
          minimum: 0
        pageSize:
          type: integer
          minimum: 0
          maximum: 200
        pageNumber:
          type: integer
          minimum: 0
        nextPage:
          type: string
    SortOrder:
      type: string
      default: ASC
      pattern: ^(ASC|DESC)$
...
    SortByRecord:
      type: string
      default: createdTimestamp
      pattern: ^(id|createdTimestamp|lastUpdatedTimestamp|status|key)$
...
  parameters:
...
    sortByRecord:
      in: query
      name: sortBy
      description: Field to sort by. createdTimestamp is chosen by default.
      schema:
        type: string
        $ref: "#/components/schemas/SortByRecord"
...
    sortOrder:
      in: query
      name: sortOrder
      description: Determines direction for resulting sorting.
      required: false
      schema:
        type: string
        $ref: "#/components/schemas/SortOrder"
    pageNumber:
      in: query
      name: pageNumber
      description: the page number to show transactions for.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    pageSize:
      in: query
      name: pageSize
      description: the number of transactions to retrieve.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50